+++ src/mip-story/mip-story.js
@@ -20,6 +20,7 @@ define(function (require) {
 
     require('./mip-story-view');
     require('./mip-story-layer');
+    var Audio = require('./audio');
     var ShareLayer = require('./mip-story-share');
     var HintLayer = require('./mip-story-hint');
     var BookEnd = require('./mip-story-bookend');
@@ -63,17 +64,11 @@ define(function (require) {
     };
 
     MIPStory.prototype.initAudio = function () {
-        var audioSrc = this.element.getAttribute('background-audio');
-        if (audioSrc) {
-            var audioEl = document.createElement('audio');
-            audioEl.setAttribute('src', audioSrc);
-            audioEl.setAttribute('preload', 'auto');
-            audioEl.setAttribute('loop', '');
-            audioEl.setAttribute('autoplay', '');
-            audioEl.setAttribute('muted', '');
-            audioEl.style.disply = 'hidden';
-            this.element.appendChild(audioEl);
-        }
+        var au = this.element.getAttribute('background-audio');
+        if (au) {
+            this.audio = new Audio();
+            au && this.audio.build(this.element, au);
+        }        
     };
 
     MIPStory.prototype.initShare = function () {
@@ -239,10 +234,11 @@ define(function (require) {
         }
         var currentEle = storyViews[this.currentIndex];
         var preEle = storyViews[this.preInex];
+        var needInitAudio = this.audio ? false : true;
         if (this.currentIndex !== this.preInex) {
-            preEle.customElement.setActive(false, this.muted);
+            preEle.customElement.setActive(false, this.muted, needInitAudio);
         }
-        currentEle.customElement.setActive(true, this.muted);
+        currentEle.customElement.setActive(true, this.muted, needInitAudio);
         this.progress.updateProgress(this.currentIndex, data.status);
         this.preInex = this.currentIndex;
